requirements.txt
Flask==2.0.1
flask-mail==0.9.1
python-dotenv==0.19.0
flask-cors==3.0.10
gunicorn==20.1.0
Werkzeug==2.0.1 